delete alert and messaging

+++ src/pages/courses/CoursesPage.js
@@ -37,11 +37,13 @@ class CoursesPage extends React.Component {
   };
 
   handleDeleteCourse = async (course) => {
-    toast.success('Course deleted');
-    try {
-      await this.props.actions.deleteCourse(course, this.props.accessToken);
-    } catch (error) {
-      toast.error('Delete failed. ' + error.message, { autoClose: false });
+    if (window.confirm('Are you sure you want to delete this course?')) {
+      toast.success('Course deleted');
+      try {
+        await this.props.actions.deleteCourse(course, this.props.accessToken);
+      } catch (error) {
+        toast.error('Delete failed. ' + error.message, { autoClose: false });
+      }
     }
   };
